+++ app/scripts/metamask-controller.js
@@ -675,6 +675,10 @@ module.exports = class MetamaskController extends EventEmitter {
     })
   }
 
+  getCurrentNetwork = () => {
+    return this.networkController.store.getState().network
+  }
+
   /**
    * Collects all the information that we want to share
    * with the mobile client for syncing purposes